+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.8.0+34 (in progress)
 
 - move logout to settings page.
+- add export button in settings page.
 
 ## 2.7.0+33